PaymentServiceProject/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
environ==1.0
pytest==7.1.2
requests==2.28.1
